+++ src/index.js
@@ -7,6 +7,7 @@ import authRouter from './routes/auth.js';
 import profileRouter from './routes/profile.js';
 import freeDictionaryRouter from './routes/freedictionary.js';
 import cardsRouter from './routes/card.js';
+import practicesRouter from './routes/practice.js';
 import { protect } from './middlewares/auth.js';
 import getLogger from './utils/logger.js';
 
@@ -29,6 +30,7 @@ app.use('/auth', authRouter);
 app.use('/profile', protect, profileRouter);
 app.use('/freedictionary', protect, freeDictionaryRouter);
 app.use('/cards', protect, cardsRouter);
+app.use('/practices', protect, practicesRouter);
 
 app.listen(port, () => {
   logger.info(`[index] app start, listening port ${port} - ${process.env.NODE_ENV}`);
