+++ src/routes/card.js
@@ -4,7 +4,6 @@ import {
   getCard,
   updateCard,
   deleteCard,
-  getCards,
 } from '../controllers/card.js';
 import { createCardValidator } from '../validators/card.js';
 import { runValidation } from '../validators/index.js';
@@ -15,6 +14,5 @@ cardsRouter.post('/', createCardValidator, runValidation, createCard);
 cardsRouter.get('/:id', getCard);
 cardsRouter.patch('/:id', runValidation, updateCard);
 cardsRouter.delete('/:id', deleteCard);
-cardsRouter.get('/', getCards);
 
 export default cardsRouter;
