+++ src/validators/index.js
@@ -1,10 +1,13 @@
+import getLogger from '../utils/logger.js';
 import { validationResult } from 'express-validator';
+
+const logger = getLogger();
 
 export const runValidation = (req, res, next) => {
   const errors = validationResult(req);
 
   if (!errors.isEmpty()) {
-    console.log('VALIDATION ERRORS', errors);
+    logger.error(`[validators/index/runValidation] errors = ${JSON.stringify(errors)}`);
     return res.status(422).json({
       error: errors.array()[0].msg,
     });
