+++ src/validators/practice.js
@@ -1,0 +1,9 @@
+import { check } from 'express-validator';
+
+const getPracticeValidator = [
+  check('practiceType').isIn(['daily']).withMessage('Invalid practice type'),
+];
+
+export {
+  getPracticeValidator,
+}
